testsuites/psxtests/*/init.c: Address type limit warnings
GCC generates type limit warnings when comparison is always true or
always false due to the limited range of the data type, This was in a
termios structure. For most of the cases in RTEMS, this was when
a variable of an unsigned type was compared to be < or <= 0.

+++ testsuites/psxtests/psxconfig01/init.c
@@ -461,7 +461,7 @@ static rtems_task Init(rtems_task_argument argument)
     };
     char *path = get_posix_name('M', 'S', 'Q', i);
     mqd_t mq = mq_open(path, oflag, mode, &attr);
-    rtems_test_assert(mq >= 0);
+    rtems_test_assert((int) mq >= 0);
   }
   rtems_resource_snapshot_take(&snapshot);
   rtems_test_assert(
